+++ apps/app/src/main.ts
@@ -3,6 +3,7 @@ import { createProvider, LocalService, createModules } from '@vtj/web';
 import router from './router';
 import App from './App.vue';
 import './style/index.scss';
+import { name } from '../package.json';
 
 const app = createApp(App);
 const service = new LocalService();
@@ -16,6 +17,9 @@ const { provider, onReady } = createProvider({
     Vue: () => import('vue'),
     VueRouter: () => import('vue-router'),
     ElementPlus: () => import('element-plus')
+  },
+  project: {
+    id: name
   }
 });