+++ packages/designer/src/components/hooks/useBinder.ts
@@ -43,7 +43,7 @@ export function useBinder(
 
     // 组件props
     const props = current.props.map((n) =>
-      typeof n === 'string' ? `this.$props.${n}` : `this.$props.${n.name}`
+      typeof n === 'string' ? `this.props.${n}` : `this.props.${n.name}`
     );
     if (props.length) {
       opts.push({ title: '属性', items: props });