+++ apps/plugin/CHANGELOG.md
@@ -3,6 +3,27 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.2](https://gitee.com/newgateway/vtj/compare/vtj-block-plugin@0.8.1...vtj-block-plugin@0.8.2) (2024-04-22)
+
+**Note:** Version bump only for package vtj-block-plugin
+
+
+
+
+
+## [0.8.1](https://gitee.com/newgateway/vtj/compare/vtj-block-plugin@0.2.0...vtj-block-plugin@0.8.1) (2024-04-22)
+
+
+### Bug Fixes
+
+* 🐛 cli templates deps ([e509c3f](https://gitee.com/newgateway/vtj/commits/e509c3fee7d360654944dab79b482dee133da638))
+* 🐛 version ([cf6e4c9](https://gitee.com/newgateway/vtj/commits/cf6e4c90de2ed12cb193c2c9929c16a8f26fa72d))
+
+
+
+
+
+
 # 0.2.0 (2024-04-22)