+++ dev/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.2](https://gitee.com/newgateway/vtj/compare/dev-web@0.8.1...dev-web@0.8.2) (2024-04-22)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
+## [0.8.1](https://gitee.com/newgateway/vtj/compare/dev-web@0.8.0...dev-web@0.8.1) (2024-04-22)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
+
 # [0.8.0](https://gitee.com/newgateway/vtj/compare/dev-web@0.7.34...dev-web@0.8.0) (2024-04-22)